refactor(core): split order item option resolution
- move size, milk, temperature, shot, and quantity resolution into a focused helper module
- keep orderItems centered on menu lookup, item resolution, and quote aggregation
- reduce orderItems.ts from 296 to 176 LOC and density from 0.17 to 0.13
